refactor(sdk/analyze): externalize subagent_tree + context_delta tests
Apply the patterns_tests.rs / hotspots_tests.rs convention to the two
remaining large analyze files with a clean single inline test block:

  subagent_tree.rs  1277 -> 826 code  (tests -> subagent_tree_tests.rs)
  context_delta.rs  1077 -> 658 code  (tests -> context_delta_tests.rs)

Tests moved verbatim, module name kept `tests` so paths are unchanged.
All tests pass, clippy clean. Pure relocation.
